+++ RELEASE-NOTES.md
@@ -6,10 +6,11 @@ kount-access-dotnet-sdk 4.0.0
 * default API version is changed from 2.1 to 4.0
 * get Kount Access data for session based on what was requested in the info flag
 * get devices that belong to a uniq user
+* get the uniq users that belong to a device
 
 ### Improvements
 * updated examples
-* logging each requests
+* logging each request
 * added more unit tests
 * changed Accept HTTP header value from text/json to application/json